// apps/core/src/enums/pdfVariable.enum.ts
export enum PdfVarKey {
  // Header / Master
  WATER_MARK = "WATER_MARK",
  COMPANY_NAME = "COMPANY_NAME",
  CC_ADDRESS_LINE_1 = "CC_ADDRESS_LINE_1",
  CC_ADDRESS_LINE_2 = "CC_ADDRESS_LINE_2",
  CC_PHONE_1 = "CC_PHONE_1",
  CC_PHONE_2 = "CC_PHONE_2",
  CC_EMAIL = "CC_EMAIL",

  // Patient / summary
  TITLE = "TITLE",
  NAME = "NAME",
  AGE = "AGE",
  SEX = "SEX",
  REF_NO = "REF_NO",
  CONTACT_NO = "CONTACT_NO",
  EMAIL = "EMAIL",
  ADDRESS = "ADDRESS",
  PATIENT_HISTORY = "PATIENT_HISTORY",
  PATIENT_ID = "PATIENT_ID",
  VISIT_NO = "VISIT_NO",
  DATE = "DATE",
  TIME = "TIME",
  VISIT_TYPE = "VISIT_TYPE",
  DELIVERY_MODE = "DELIVERY_MODE",

  // Table (line items)
  SL_NO = "SL_NO",
  TABLE_DOCTOR = "TABLE_DOCTOR",
  DATE_TIME = "DATE_TIME",
  ITEM_NAME = "ITEM_NAME",
  BATCH = "BATCH",
  EXP_DATE = "EXP_DATE",
  BASE_RATE = "BASE_RATE",
  QUANTITY = "QUANTITY",
  TAX = "TAX",
  DISCOUNT = "DISCOUNT",
  CO_PAY = "CO_PAY",
  VIP = "VIP",
  NET_AMOUNT = "NET_AMOUNT",
  OTHER_CHARGE = "OTHER_CHARGE",
  GROSS_AMOUNT = "GROSS_AMOUNT",
  SUB_TOTAL = "SUB_TOTAL",

  // Amount summary
  TOTAL_GROSS = "TOTAL_GROSS",
  TOTAL_NET = "TOTAL_NET",
  TOTAL_DISCOUNT = "TOTAL_DISCOUNT",
  TOTAL_TAX = "TOTAL_TAX",
  TOTAL_PAID = "TOTAL_PAID",
  REFUND = "REFUND",
  REFUNDED = "REFUNDED",
  DUE = "DUE",
  TOTAL_CO_PAY = "TOTAL_CO_PAY",

  // Transaction summary
  BILLED_BY = "BILLED_BY",
  REPORT_URL = "REPORT_URL",
  TRANSACTION = "TRANSACTION",
  AMOUNT_IN_WORD = "AMOUNT_IN_WORD",
  USER_LOGIN = "USER_LOGIN",
  USER_PASSWORD = "USER_PASSWORD",

  // Footer
  SIGNATURE = "SIGNATURE",
  SIGNATURE_DOCTOR = "SIGNATURE_DOCTOR",
  DOCTOR_DESIGNATION = "DOCTOR_DESIGNATION",
  DOCTOR_REG_NO = "DOCTOR_REG_NO",
  FOOTER_MESSAGE = "FOOTER_MESSAGE",
  FOOTER_EMAIL = "FOOTER_EMAIL",
}